cmd/internal,cmd/link: add support for calling SDYNIMPORT on mips64
Add support for calling SDYNIMPORT symbols on mips64 - when used with external linking,
it is expected that the caller will load register 25 (t9) with the address of the
external symbol using an offset from the global pointer (gp) register.

In order to achieve this, generate trampolines that have the correct calling convention
and appropriate relocations. It is expected that the caller will have restored the gp
register to the value initially set by the dynamic linker and then restore it to Go's
value following the call.